refactor: Update step2 package name
