+++ src/test/kotlin/step2_string_calculator/CalculatorServiceTest.kt
@@ -1,4 +1,4 @@
-package string_calculator
+package step2_string_calculator
 
 import io.kotest.assertions.throwables.shouldThrow
 import io.kotest.core.spec.style.StringSpec
